// kinowatch/src/store/action-types.ts
/* eslint-disable */
export enum ActionType {
  FETCH_INITIAL_DATA = 'FETCH_INITIAL_DATA',
  SET_IS_LOADED = 'SET_IS_LOADED',
  SET_MY_PROFILE = 'SET_MY_PROFILE',
  SET_MY_TRACKS = 'SET_MY_TRACKS',
  SET_MY_PLAYLISTS = 'SET_MY_PLAYLISTS',
  ADD_PLAYLIST = 'ADD_PLAYLIST',
  ADD_TRACKLIST = 'ADD_TRACKLIST',
  SET_RECOMENDATION = 'SET_RECOMENDATION',
  SET_CURRENT_TRACK = 'SET_CURRENT_TRACK',
  SET_CURRENT_TRACKLIST = 'SET_CURRENT_TRACKLIST',
  SET_PROGRESS = 'SET_PROGRESS',
  SET_BUFFER_TIME = 'SET_BUFFER_TIME',
  SET_IS_PLAYNING = 'SET_IS_PLAYNING',
  SET_IS_SHUFFLE = 'SET_IS_SHUFFLE',
  CHANGE_TO_PREV_TRACK = 'CHANGE_TO_PREV_TRACK',
  CHANGE_TO_NEXT_TRACK = 'CHANGE_TO_NEXT_TRACK',
}